+++ examples/example.c
@@ -317,7 +317,7 @@ main(int argc, char **argv)
         sentry_set_context("runtime", context);
 
         sentry_value_t user = sentry_value_new_object();
-        sentry_value_set_by_key(user, "id", sentry_value_new_int32(42));
+        sentry_value_set_by_key(user, "id", sentry_value_new_string("42"));
         sentry_value_set_by_key(
             user, "username", sentry_value_new_string("some_name"));
         sentry_set_user(user);