+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Use `propagation_context` as the single source of `trace_id` for spans and events. ([#1200](https://github.com/getsentry/sentry-native/pull/1200))
 
+**Features**:
+
+- Add `sentry_value_new_user(id, username, email, ip_address)` function to avoid ambiguous user-context-keys. ([#1228](https://github.com/getsentry/sentry-native/pull/1228))
+
 **Fixes**:
 
 - Remove compile-time check for the `libcurl` feature `AsynchDNS`. ([#1206](https://github.com/getsentry/sentry-native/pull/1206))